# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pending-requests": {
      "get": {
        "description": "Eventmanager views all client event hosting requests.",
        "operationId": "PendingRequestsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Pending",
                "Approved",
                "Rejected"
              ],
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending requests"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get all event requests",
        "tags": [
          "Pending Requests"
        ]
      },
      "post": {
        "description": "Client submits a request to host an event.",
        "operationId": "PendingRequestsController_create",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "superuser | client | eventmanager",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePendingRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Request submitted"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Submit new event hosting request",
        "tags": [
          "Pending Requests"
        ]
      }
    },
    "/pending-requests/{id}": {
      "get": {
        "operationId": "PendingRequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Request ID (e.g. pr1)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager | client",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request details"
          },
          "404": {
            "description": "Request not found"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Get request by ID",
        "tags": [
          "Pending Requests"
        ]
      },
      "delete": {
        "operationId": "PendingRequestsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Request ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request deleted"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Delete event request",
        "tags": [
          "Pending Requests"
        ]
      }
    },
    "/pending-requests/{id}/review": {
      "patch": {
        "description": "Eventmanager reviews a pending request. Approval requires managerId.",
        "operationId": "PendingRequestsController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Request ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Must be: superuser | eventmanager",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eventmanager"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewPendingRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request reviewed"
          },
          "400": {
            "description": "Already reviewed or missing managerId"
          }
        },
        "security": [
          {
            "role-header": []
          }
        ],
        "summary": "Approve or reject an event request",
        "tags": [
          "Pending Requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "WEvents API",
...
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role-header": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "User role (superuser | eventmanager | client | osc | attendee)"
      }
    },
    "schemas": {
...
      "CreatePendingRequestDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tech Innovation Summit 2026",
            "description": "Proposed event name"
          },
          "location": {
            "type": "string",
            "example": "Pune, Maharashtra",
            "description": "Proposed event location"
          },
          "expected": {
            "type": "number",
            "example": 500,
            "description": "Expected number of attendees"
          },
          "client": {
            "type": "string",
            "example": "Priya Sharma",
            "description": "Client / requester name"
          },
          "clientEmail": {
            "type": "string",
            "example": "priya.s@techcorp.com",
            "description": "Client email"
          },
          "date": {
            "type": "string",
            "example": "2026-05-15",
            "description": "Desired event date (YYYY-MM-DD)"
          },
          "description": {
            "type": "string",
            "example": "Annual technology conference featuring AI and blockchain workshops",
            "description": "Event description"
          },
          "type": {
            "type": "string",
            "example": "In-Person",
            "enum": [
              "In-Person",
              "Virtual",
              "Hybrid"
            ],
            "description": "Event type"
          }
        },
        "required": [
          "name",
          "location",
          "expected",
          "client",
          "clientEmail",
          "date",
          "description",
          "type"
        ]
      },
      "ReviewPendingRequestDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "example": "approved",
            "enum": [
              "approved",
              "rejected"
            ],
            "description": "Approval decision"
          },
          "managerId": {
            "type": "string",
            "example": "u1",
            "description": "Manager ID to assign (required when approving)"
          },
          "managerName": {
            "type": "string",
            "example": "Rajesh Kumar",
            "description": "Manager name"
          },
          "rejectionReason": {
            "type": "string",
            "example": "Budget constraints",
            "description": "Reason for rejection"
          }
        },
        "required": [
          "decision"
        ]
      },
...
    }
  }
}